+++ libslide/src/parser/expression_parser.rs
@@ -3,6 +3,7 @@ use crate::grammar::*;
 use crate::scanner::types::{Token, TokenType};
 use crate::utils::{hash, PeekIter, StringUtils};
 
+use rug::Float;
 use std::collections::HashMap;
 use std::rc::Rc;
 
@@ -79,7 +80,7 @@ impl Parser<Stmt> for ExpressionParser {
         parsed
     }
 
-    fn parse_float(&mut self, f: f64) -> Self::Expr {
+    fn parse_float(&mut self, f: Float) -> Self::Expr {
         Self::Expr::Const(f)
     }
